# docs/swagger.json
{"swagger":"2.0","info":{"description":"Backend API for VidyarthiDesk ERP Software","title":"VidyarthiDesk API","termsOfService":"http://swagger.io/terms/","contact":{"name":"API Support","email":"vaishakh@vaishakhgk.com"},"license":{"name":"GPL-3.0","url":"https://www.gnu.org/licenses/gpl-3.0.html"},"version":"1.0"},"host":"localhost:9000","basePath":"/api","paths":{"/auth/login":{"post":{"description":"Authenticate user with email and password. Returns JWT access token and refresh token stored in session.","consumes":["application/json"],"produces":["application/json"],"tags":["Authentication"],"summary":"User login","parameters":[{"description":"Login credentials","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.LoginReq"}}],"responses":{"200":{"description":"Successfully authenticated","schema":{"$ref":"#/definitions/dto.LoginRes"}},"400":{"description":"Invalid request body or validation error","schema":{"$ref":"#/definitions/dto.ErrorRes"}},"401":{"description":"Invalid email or password","schema":{"$ref":"#/definitions/dto.ErrorRes"}},"500":{"description":"Internal server error","schema":{"$ref":"#/definitions/dto.ErrorRes"}}}}},"/auth/me":{"get":{"security":[{"BearerAuth":[]}],"description":"Get the authenticated user's information from JWT token","produces":["application/json"],"tags":["Authentication"],"summary":"Get current user","responses":{"200":{"description":"Current user information","schema":{"$ref":"#/definitions/dto.UserRes"}},"401":{"description":"user not found","schema":{"$ref":"#/definitions/dto.ErrorRes"}},"500":{"description":"Internal server error","schema":{"$ref":"#/definitions/dto.ErrorRes"}}}}},"/auth/refresh":{"post":{"description":"Generate new access token and refresh token using existing refresh token from session. Old refresh token is revoked.","produces":["application/json"],"tags":["Authentication"],"summary":"Refresh access token","responses":{"200":{"description":"New tokens generated successfully","schema":{"$ref":"#/definitions/dto.LoginRes"}},"401":{"description":"No refresh token in session, token expired, or token revoked","schema":{"$ref":"#/definitions/dto.ErrorRes"}},"500":{"description":"Internal server error","schema":{"$ref":"#/definitions/dto.ErrorRes"}}}}}},"definitions":{"dto.ErrorRes":{"type":"object","properties":{"error":{"type":"string","example":"Invalid credentials"},"message":{"type":"string","example":"Email or password is incorrect"}}},"dto.LoginReq":{"description":"Login request payload","type":"object","required":["email","password"],"properties":{"email":{"type":"string","maxLength":50,"minLength":3,"example":"admin"},"password":{"type":"string","minLength":6,"example":"password"}}},"dto.LoginRes":{"description":"Login response with tokens","type":"object","properties":{"authToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"refreshToken":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"}}},"dto.UserRes":{"description":"User information structure","type":"object","properties":{"createdAt":{"type":"string","example":"2024-01-01T00:00:00Z"},"email":{"type":"string","example":"user@example.com"},"id":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"name":{"type":"string","example":"John Doe"},"updatedAt":{"type":"string","example":"2024-01-15T10:30:00Z"}}}},"securityDefinitions":{"BearerAuth":{"description":"Type \"Bearer\" followed by a space and JWT token.","type":"apiKey","name":"Authorization","in":"header"}}}